Moved several classes into same package.
