+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/AttachmentStatusException.java
@@ -1,7 +1,7 @@
 /**************************************************************************
  * (C) 2019-2024 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.exception;
+package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper;
 
 import com.sap.cds.feature.attachments.utilities.AttachmentErrorStatuses;
 import com.sap.cds.services.ErrorStatus;
@@ -9,16 +9,18 @@ import com.sap.cds.services.utils.ErrorStatusException;
 
 public class AttachmentStatusException extends ErrorStatusException {
 
+	private static final long serialVersionUID = 743951900719373778L;
+
 	private AttachmentStatusException(ErrorStatus errorStatus, String message) {
 		super(errorStatus, message);
 	}
 
-	public static AttachmentStatusException getNotCleanException() {
+	static AttachmentStatusException getNotCleanException() {
 		return new AttachmentStatusException(AttachmentErrorStatuses.NOT_CLEAN,
 				AttachmentErrorStatuses.NOT_CLEAN.getDescription());
 	}
 
-	public static AttachmentStatusException getNotScannedException() {
+	static AttachmentStatusException getNotScannedException() {
 		return new AttachmentStatusException(AttachmentErrorStatuses.NOT_SCANNED,
 				AttachmentErrorStatuses.NOT_SCANNED.getDescription());
 	}
