+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/AttachmentStatusValidator.java
@@ -1,10 +1,9 @@
 /**************************************************************************
  * (C) 2019-2024 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.validator;
+package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper;
 
 import com.sap.cds.feature.attachments.generated.cds4j.sap.attachments.StatusCode;
-import com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.exception.AttachmentStatusException;
 
 public class AttachmentStatusValidator {
