+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/BeforeReadItemsModifier.java
@@ -1,7 +1,7 @@
 /**************************************************************************
  * (C) 2019-2024 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.modifier;
+package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper;
 
 import java.util.ArrayList;
 import java.util.List;
