+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/AttachmentStatusExceptionTest.java
@@ -1,9 +1,10 @@
-package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.exception;
+package com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
+import com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.AttachmentStatusException;
 import com.sap.cds.feature.attachments.utilities.AttachmentErrorStatuses;
 
 class AttachmentStatusExceptionTest {
